+++ database/migrations/2025_06_30_163700_create_bookings_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('bookings', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('trip_id')->constrained()->onDelete('cascade');
-            $table->char('booking_code', 10)->unique();
-            $table->string('booking_reference', 20)->unique();
-            $table->string('passenger_name');
-            $table->string('passenger_phone', 20);
-            $table->string('passenger_email')->nullable();
-            $table->enum('status', ['booked', 'cancelled'])->default('booked');
-            $table->decimal('total_amount', 10, 2);
-            $table->enum('payment_status', ['pending', 'paid', 'failed', 'refunded'])->default('pending');
-            $table->string('cancellation_reason')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('bookings')) {
+            Schema::create('bookings', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('trip_id')->constrained()->onDelete('cascade');
+                $table->char('booking_code', 10)->unique();
+                $table->string('booking_reference', 20)->unique();
+                $table->string('passenger_name');
+                $table->string('passenger_phone', 20);
+                $table->string('passenger_email')->nullable();
+                $table->enum('status', ['booked', 'cancelled'])->default('booked');
+                $table->decimal('total_amount', 10, 2);
+                $table->enum('payment_status', ['pending', 'paid', 'failed', 'refunded'])->default('pending');
+                $table->string('cancellation_reason')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
