+++ database/migrations/2025_06_30_163720_create_payments_table.php
@@ -11,17 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('payments', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('booking_id')->constrained()->onDelete('cascade');
-            $table->decimal('amount', 10, 2);
-            $table->enum('method', ['eSewa', 'Khalti', 'Cash']);
-            $table->enum('payment_status', ['pending', 'success', 'failed', 'refunded'])->default('pending');
-            $table->string('transaction_id', 100)->nullable();
-            $table->datetime('payment_time')->nullable();
-            $table->json('response_log')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('payments')) {
+            Schema::create('payments', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('booking_id')->constrained()->onDelete('cascade');
+                $table->decimal('amount', 10, 2);
+                $table->enum('method', ['eSewa', 'Khalti', 'Cash']);
+                $table->enum('payment_status', ['pending', 'success', 'failed', 'refunded'])->default('pending');
+                $table->string('transaction_id', 100)->nullable();
+                $table->datetime('payment_time')->nullable();
+                $table->json('response_log')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
